Display both fractions and decimals in result
Changes the format of results displayed to the user.

A decimal will always be displayed unless it is the same as
the user's query or the result is an integer.

In the case of the result being an integer, then just this value
will be displayed.

A fraction will only be displayed if it does not match the user's
query and all the operations produced rational numbers.

In the case that just a decimal is displayed, if the decimal
is not exact, then an 'Approximately Equal To' symbol will be
displayed in front of the number.

In the case that both a decimal and fraction are displayed
(the decimal always being placed after the fraction), if
the decimal is exact, then an equals sign will be placed after
the fraction, otherwise an 'Approximately Equal To' symbol will
be used.